# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_media_item_catalog__entry_id__items_media_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 300,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_media_item_catalog__entry_id__items_media_post",
        "type": "object"
      },
      "Body_upload_media_me_library_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 300,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_media_me_library_upload_post",
        "type": "object"
      },
...
      "GenerateRequest": {
        "description": "Ask the AI layer to generate an image + text into the user's Agent storage (AC51).",
        "properties": {
          "prompt": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemKind": {
        "description": "The kind of a first-class entry item (AC50).",
        "enum": [
          "text",
          "image",
          "video"
        ],
        "title": "ItemKind",
        "type": "string"
      },
...
      "MediaItem": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "entry_id": {
            "title": "Entry Id",
            "type": "integer"
          },
          "entry_title": {
            "title": "Entry Title",
            "type": "string"
          },
          "object_key": {
            "title": "Object Key",
            "type": "string"
          }
        },
        "required": [
          "object_key",
          "content_type",
          "entry_id",
          "entry_title"
        ],
        "title": "MediaItem",
        "type": "object"
      },
...
      "TextItemCreate": {
        "properties": {
          "order": {
            "default": 0,
            "title": "Order",
            "type": "integer"
          },
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "title": {
            "default": "",
            "maxLength": 300,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextItemCreate",
        "type": "object"
      },
...
      "UserAssetOut": {
        "description": "A user's own stored asset (AC51) — uploaded (``local``) or generated (``agent``).",
        "properties": {
          "content_type": {
            "default": "",
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/ItemKind"
          },
          "object_key": {
            "default": "",
            "title": "Object Key",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/UserAssetSource"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "kind"
        ],
        "title": "UserAssetOut",
        "type": "object"
      },
      "UserAssetSource": {
        "description": "Origin of a user-storage asset (AC51): user upload, or AI-service generation.",
        "enum": [
          "local",
          "agent"
        ],
        "title": "UserAssetSource",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Walsh Content Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/catalog/{entry_id}/items/media": {
      "post": {
        "description": "Add an image or video media item to an entry (AC50). Stored in the provider's media store.",
        "operationId": "upload_media_item_catalog__entry_id__items_media_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_item_catalog__entry_id__items_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Upload Media Item Catalog  Entry Id  Items Media Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Media Item",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/me/library": {
      "get": {
        "description": "List the caller's own assets, optionally filtered by source (Local / Agent). Owner-scoped.",
        "operationId": "list_media_me_library_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserAssetSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserAssetOut"
                  },
                  "title": "Response List Media Me Library Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Media",
        "tags": [
          "me-library"
        ]
      }
    },
    "/me/library/generate": {
      "post": {
        "description": "Generate an image + text into your Agent storage (AC51). Deterministic PoC stub — image +\ntext only; no generated video/animation (that comes from sprites).",
        "operationId": "generate_me_library_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserAssetOut"
                  },
                  "title": "Response Generate Me Library Generate Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate",
        "tags": [
          "me-library"
        ]
      }
    },
    "/me/library/text": {
      "post": {
        "description": "Save a text snippet into your Local storage (AC51).",
        "operationId": "add_text_me_library_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAssetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Text",
        "tags": [
          "me-library"
        ]
      }
    },
    "/me/library/upload": {
      "post": {
        "description": "Upload an image or video into your Local storage (AC51).",
        "operationId": "upload_media_me_library_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_me_library_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAssetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Media",
        "tags": [
          "me-library"
        ]
      }
    },
    "/me/media": {
      "get": {
        "description": "Every image/asset across this provider's catalog entries, in one place (AC29).",
        "operationId": "media_library_me_media_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaItem"
                  },
                  "title": "Response Media Library Me Media Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Media Library",
        "tags": [
          "provider"
        ]
      }
    },
...
  }
}